requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
diff-match-patch==20230430
Django==5.1.1
django-crispy-forms==2.3
django-import-export==4.1.1
djangorestframework==3.15.2
idna==3.8
isodate==0.6.1
lxml==5.3.0
pillow==10.4.0
platformdirs==4.3.2
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
urllib3==2.2.2
zeep==4.2.1
